fix 'float' object cannot be interpreted as an index error

+++ epson_printer/epsonprinter.py
@@ -123,7 +123,7 @@ class PrintableImage:
         extra_pixels = np.ones((extra_rows, w), dtype=bool)
         pixels = np.vstack((pixels, extra_pixels))
         h += extra_rows
-        nb_stripes = h / 24
+        nb_stripes = int(h / 24)
         pixels = pixels.reshape(nb_stripes, 24, w).swapaxes(1, 2).reshape(-1, 8)
 
         nh = int(w / 256)
